EmsEvent::Automate#parse_policy_parameters: always return 3 values.
Same check whether any of them is nil is done in the only caller
(EmsEvent::Automate#policy) anyway.
Previously caller could try to destructure `a, b, c = nil` which
I'm dismayed to see is not an error but sets all 3 to nil,
so it behaved the same but it was hard to understand why.