Pending test for ContainerLabelTagMapping error when used in 2 refresh workers.
This is not a low-probability race condition, it's almost certain.
Currently the cache is held for the life of a refresh worker,
so if you have 2 container providers, define an any-value mapping,
and assign such a label (with same value) to entities in both clusters,
the 2nd refresh worker to see the label will crash.